feat: added a new class Vector2Elm to deal with 2 element vectors
The class also supports addition and subtraction of two Vector2Elm,
the dot product of two Vector2Elm, the vectorial product of two
Vector2Elm, a Vector2Elm times a scalar and the absolute value of any
Vector2Elm as a static method. There is a test in the main function
to be deleted later since it is not going to be part of the program.