feat: Add lock center view to the rotation of a particular body. Pressing l alternates between lock and unlock.
refactor: Added std::shared_ptr thisbody and refBody to simplify
the code. Functions Body::Transform() moved to Vector2elm class.

fix: New method Core::multiplyDelta to limit the values of delta_.

Core new member isLocked_ to control if locked to a rotation of a body
or not.
Vector2Elm new function RotationTransform calculate position of bodies,
relative to the locked body.
Core::EventHandler modified to support the l key and to limit delta_.